requirements.txt
# Python 3.3+
beautifulsoup4>=4.4.0,<5
prettytable>=0.7.2,<1
python-debian>=0.1.23
requests>=2.2.0,<3
